ENG-296: stock qty conversion from float64 to int as cart and order supports int currently

+++ app/actors/product/decl.go
@@ -30,7 +30,7 @@ type DefaultProduct struct {
 
 	Weight float64
 
-	Qty float64
+	Qty int
 
 	Options map[string]interface{}
